chore: run prettier on repo
In my last PR, my editor initially made a bunch of formatting
changes which I had to revert. I noticed that things were quite
inconsistent with regards to quotes and general formatting across
the repo.

This PR is the result of running `prettier` with the standard
settings which normalises quotes, standardises formatting in MD
and removes spurious whitespace.